+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Added timeSlice estimation
 - Fixed normalization constant in creating sparse var matrix
+- Fixed parallel backend in Windows 
 
 # sparsevar 0.0.4